Added working links in footer
