--- react-portfolio/src/Footer.js
@@ -1,23 +1,0 @@
-import React from 'react';
-
-function Footer() {
-  return (
-    <footer>
-      <div className="footContain">
-        <div className='Icons'>
-          <ul>
-            <li><i class="fa-brands fa-square-github"></i></li>
-            <li><i class="fa-brands fa-linkedin"></i></li>
-            <li><i class="fa-brands fa-square-instagram"></i></li>
-          </ul>
-        </div>
-        <div className='Credit'>
-          <p>Background Video provided by Vecteezy</p>
-          <p>© 2023 Developed & Designed by Todd Dharni</p>
-        </div>
-      </div>
-    </footer>
-  );
-}
-
-export default Footer
